Milk Bottles Counting/requirements.txt
alabaster==0.7.13
arrow==1.3.0
astroid==3.0.3
asttokens==2.4.1
atomicwrites==1.4.1
attrs==23.2.0
autopep8==2.0.4
Babel==2.14.0
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.2.0
bleach==6.1.0
blinker==1.7.0
certifi==2022.12.7
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contourpy==1.1.1
cookiecutter==2.6.0
cryptography==42.0.4
cvzone==1.6.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.8
docstring-to-markdown==0.15
docutils==0.20.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.9.0
filterpy==1.4.5
flake8==7.0.0
Flask==3.0.2
fonttools==4.49.0
fsspec==2023.4.0
idna==3.4
imageio==2.34.0
imagesize==1.4.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.29.2
ipython==8.12.3
isort==5.13.2
itsdangerous==2.1.2
jaraco.classes==3.3.1
jedi==0.19.1
jellyfish==1.0.3
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
keyring==24.3.0
kiwisolver==1.4.5
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
mpmath==1.3.0
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.0
numpy==1.24.1
numpydoc==1.6.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycodestyle==2.11.1
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.17.2
pylint==3.0.3
pylint-venv==3.0.3
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
python-dateutil==2.8.2
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.10.0
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2024.1
PyWavelets==1.4.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
QDarkStyle==3.2.3
qstylizer==0.2.2
QtAwesome==1.3.0
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
requests==2.28.1
rich==13.7.0
rope==1.12.0
rpds-py==0.18.0
Rtree==1.2.0
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.2
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.1.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
# spyder==5.5.1
# spyder-kernels==2.5.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
text-unidecode==1.3
textdistance==4.6.1
thop==0.1.1.post2209072238
three-merge==0.1.1
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.3
# torch==2.2.0+cu118
# torchaudio==2.2.0+cu118
# torchvision==0.17.0+cu118
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.8.0
tzdata==2024.1
ujson==5.9.0
ultralytics==8.1.16
urllib3==1.26.13
watchdog==4.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
whatthepatch==1.0.5
yapf==0.40.2
zipp==3.17.0
